Add handling of media files removal
Media files SSE channels is being watched for "removal" events.
