+++ src/functions/app_init.ts
@@ -1,10 +1,16 @@
-import { getMediaFilesSse, getPlaybackSse, getPlaylistSse, getStatusSse, init as initSse } from '../functions/sse';
 import { mediaFilesStore } from '../stores/media_files';
 import { playbackStore } from '../stores/playback';
 import { apiConnectionStore } from '../stores/api_connection';
 
 import { initPlaybackHistoryWatch } from './playback_history';
 import { playlistsStore } from '../stores/playlists';
+import { filter } from 'rxjs/operators';
+import type { MediaFilesMap } from '../models/api';
+import { getMediaFilesSse, MediaFilesEvents } from './sse/media_files';
+import { getPlaybackSse } from './sse/playback';
+import { getPlaylistSse } from './sse/playlists';
+import { getStatusSse } from './sse/status';
+import { init as initSse } from './sse/initialization';
 
 export function appInit() {
   initPlaybackHistoryWatch();
@@ -42,19 +48,46 @@ export function appInit() {
     },
   });
 
-  getMediaFilesSse().subscribe(
-    mediaFiles => {
-      mediaFilesStore.update(state => {
-        return {
-          mediaFiles: {
-            ...state.mediaFiles,
-            ...mediaFiles,
-          },
-          isFetchingInProgress: false,
-        };
-      });
-    },
-  );
+  getMediaFilesSse()
+    .pipe(
+      filter(action => action.event === MediaFilesEvents.Added),
+    )
+    .subscribe(
+      action => {
+        mediaFilesStore.update(state => {
+          return {
+            mediaFiles: {
+              ...state.mediaFiles,
+              ...action.items,
+            },
+            isFetchingInProgress: false,
+          };
+        });
+      },
+    );
 
+  getMediaFilesSse()
+    .pipe(
+      filter(action => action.event === MediaFilesEvents.Removed),
+    )
+    .subscribe(
+      action => {
+        mediaFilesStore.update(state => {
+          const filteredMediaFiles: MediaFilesMap = {};
+          for (const path in state.mediaFiles) {
+            if (!!action.items[path]) continue;
+
+            Object.assign(filteredMediaFiles, { [path]: state.mediaFiles[path] });
+          }
+
+          return {
+            mediaFiles: {
+              ...filteredMediaFiles,
+            },
+            isFetchingInProgress: false,
+          };
+        });
+      },
+    );
   initSse();
 }
